+++ platform/windows/capi/include/ag_dns.h
@@ -202,6 +202,12 @@ typedef struct {
      * WARNING: may increase data usage and probability of TCP fallbacks.
      */
     bool enable_dnssec_ok;
+    /**
+     * If enabled, detect retransmitted requests and handle them using fallback upstreams only.
+     * If a retransmitted request is detected, the original request, as well as any more
+     * retransmitted requests after the first one detected, will NOT be answered.
+     */
+    bool enable_retransmission_handling;
 } ag_dnsproxy_settings;
 
 typedef struct {