+++ platform/windows/cs/Adguard.Dns/Adguard.Dns.TestApp/Program.cs
@@ -176,7 +176,8 @@ namespace Adguard.Dns.TestApp
                 },
                 Ipv6Available = false,
                 OptimisticCache = false,
-                EnableDNSSECOK = false
+                EnableDNSSECOK = false,
+                EnableRetransmissionHandling = false
             };
 
             return dnsProxySettings;